[ Chore #160705539 ] Reorganize API structure
