+++ .gitignore
@@ -2,4 +2,5 @@
 .pytest_cache/
 __pycache__/
 instance/
+.env
 env/
