+++ app/api/V1/models.py
@@ -15,3 +15,5 @@ class OrderFood:
     def current_menu(self):
         """Create a new list to store the menu items"""
         return self.menu
+
+
